CORE: Pets in Battlefields
1) Upon pet load, if the master has a battlefield, that will be propagated
to the pet and the pet will be given EFFECT_BATTLEFIELD.  This resolves
issues surrounding pets due to not being added to the battlefield.

In theory, checking for removal from a battlefield for pets should be
unnecessary, as the pet is reloaded each map change and since the master
wouldn't be in a battlefield at that time it wouldn't be readded.  This
doesn't account for all potential collisions however.

This is not a permanent solution as there's no handling for pre-existing
pets upon entering a battlefield and other such cases.  This is just a
temporary bandaid to make pets usable.